@@ -1,3 +1,5 @@
 module github.com/Xhofe/alist-proxy
 
-go 1.17
+go 1.19
+
+require github.com/alist-org/alist/v3 v3.0.0-beta.0 // indirect
